fix(houdini-svelte): circular dependencies involving plugin/kit.ts
houdini-svelte/src/plugin/kit.ts -> index.ts (not caught by madge)
houdini-svelte/src/plugin/kit.ts -> houdini-svelte/src/plugin/extract.ts
houdini-svelte/src/plugin/kit.ts -> houdini-svelte/src/plugin/extractLoadFunction.ts
