+++ packages/houdini-svelte/src/plugin/codegen/stores/fragment.ts
@@ -1,7 +1,7 @@
 import type { Document, GenerateHookInput } from 'houdini'
 import { fs, path } from 'houdini'
 
-import type { HoudiniSvelteConfig } from '../..'
+import type { HoudiniSvelteConfig } from '../../config'
 import { stores_directory, store_name } from '../../kit'
 import { store_import } from './custom'
