+++ packages/houdini-svelte/src/plugin/transforms/kit/load.ts
@@ -7,19 +7,18 @@ import { artifact_import, ensure_imports, find_insert_index } from 'houdini/vite
 import * as recast from 'recast'
 
 import { parseSvelte } from '../../extract'
-import { extract_load_function } from '../../extractLoadFunction'
-import type { HoudiniRouteScript } from '../../kit'
+import { extract_load_function, type HoudiniRouteScript } from '../../extractLoadFunction'
 import {
 	is_layout,
 	is_route,
 	is_route_script,
 	layout_query_path,
 	page_query_path,
-	plugin_config,
 	route_data_path,
 	route_page_path,
 	store_import_path,
 } from '../../kit'
+import { plugin_config } from '../../config'
 import {
 	houdini_afterLoad_fn,
 	houdini_before_load_fn,
