+++ packages/houdini-svelte/src/runtime/stores/query.ts
@@ -13,7 +13,7 @@ import { ArtifactKind, CachePolicy, CompiledQueryKind } from '$houdini/runtime/l
 import type { LoadEvent } from '@sveltejs/kit'
 import { get } from 'svelte/store'
 
-import type { HoudiniSvelteConfig } from '../../plugin'
+import type { HoudiniSvelteConfig } from '../../plugin/config'
 import type { PluginArtifactData } from '../../plugin/artifactData'
 import { clientStarted, isBrowser } from '../adapter'
 import { initClient } from '../client'
